Silenced a couple of warnings in VSRip, updated Authors
git-svn-id: https://mpc-hc.svn.sourceforge.net/svnroot/mpc-hc/trunk@2418 10f7b99b-c216-0410-bff0-8a66a9350fd8

+++ src/apps/VSRip/VSRipPGCDlg.cpp
@@ -110,7 +110,7 @@ void CVSRipPGCDlg::SetupPGCList()
 
 	m_pgclist.ResetContent();
 
-	for(int i = 0; i < m_rd.pgcs.GetCount(); i++)
+	for(int i = 0; i < (int)m_rd.pgcs.GetCount(); i++)
 	{
 		CString str;
 		str.Format(_T("PGC %d"), i+1);
@@ -189,7 +189,7 @@ void CVSRipPGCDlg::SetupVCList()
 
 	CAtlArray<vc_t>& vca =( m_rd.pgcs[m_rd.iSelPGC].angles[m_rd.pgcs[m_rd.iSelPGC].iSelAngle]);
 
-	for(int i = 0; i < vca.GetCount(); i++)
+	for(int i = 0; i < (int)vca.GetCount(); i++)
 	{
 		CString str;
 		str.Format(_T("V%02d C%02d"), vca[i].vob, vca[i].cell);
